+++ connect/common/models.py
@@ -11,7 +11,7 @@ import stripe
 from celery import current_app
 from django.conf import settings
 from django.db import models
-from django.db.models import Q, Sum
+from django.db.models import Sum
 from django.template.loader import render_to_string
 from django.utils import timezone, translation
 from django.utils.translation import activate, ugettext_lazy as _
@@ -21,12 +21,15 @@ from model_utils import FieldTracker
 
 from connect import billing
 from connect.api.v1.internal.flows.flows_rest_client import FlowsRESTClient
-from connect.api.v1.internal.intelligence.intelligence_rest_client import \
-    IntelligenceRESTClient
+from connect.api.v1.internal.intelligence.intelligence_rest_client import (
+    IntelligenceRESTClient,
+)
 from connect.authentication.models import User
 from connect.billing.gateways.stripe_gateway import StripeGateway
-from connect.common.exceptions import (OrganizationAuthorizationException,
-                                       ProjectAuthorizationException)
+from connect.common.exceptions import (
+    OrganizationAuthorizationException,
+    ProjectAuthorizationException,
+)
 from connect.common.gateways.rocket_gateway import Rocket
 from connect.common.helpers import send_mass_html_mail
 from connect.internals.event_driven.producer.rabbitmq_publisher import RabbitmqPublisher
@@ -298,11 +301,11 @@ class Organization(models.Model):
 
 
 class OrganizationLevelRole(Enum):
-    NOTHING, VIEWER, CONTRIBUTOR, ADMIN, FINANCIAL, SUPPORT = list(range(6))
+    NOTHING, VIEWER, CONTRIBUTOR, ADMIN, FINANCIAL, SUPPORT, MARKETING = list(range(7))
 
 
 class OrganizationRole(Enum):
-    NOT_SETTED, VIEWER, CONTRIBUTOR, ADMIN, FINANCIAL, SUPPORT = list(range(6))
+    NOT_SETTED, VIEWER, CONTRIBUTOR, ADMIN, FINANCIAL, SUPPORT, MARKETING = list(range(7))
 
 
 class OrganizationAuthorization(models.Model):
@@ -318,6 +321,7 @@ class OrganizationAuthorization(models.Model):
         (OrganizationRole.VIEWER.value, _("viewer")),
         (OrganizationRole.FINANCIAL.value, _("financial")),
         (OrganizationRole.SUPPORT.value, _("support")),
+        (OrganizationRole.MARKETING.value, _("marketing")),
     ]
 
     uuid = models.UUIDField(
@@ -356,6 +360,9 @@ class OrganizationAuthorization(models.Model):
         if self.role == OrganizationRole.SUPPORT.value:
             return OrganizationLevelRole.SUPPORT.value
 
+        if self.role == OrganizationRole.MARKETING.value:
+            return OrganizationLevelRole.MARKETING.value
+
     @property
     def can_read(self):
         return self.level in [
@@ -364,6 +371,7 @@ class OrganizationAuthorization(models.Model):
             OrganizationLevelRole.ADMIN.value,
             OrganizationLevelRole.VIEWER.value,
             OrganizationLevelRole.SUPPORT.value,
+            OrganizationLevelRole.MARKETING.value,
         ]
 
     @property
@@ -372,6 +380,7 @@ class OrganizationAuthorization(models.Model):
             OrganizationLevelRole.CONTRIBUTOR.value,
             OrganizationLevelRole.ADMIN.value,
             OrganizationLevelRole.SUPPORT.value,
+            OrganizationLevelRole.MARKETING.value,
         ]
 
     @property
@@ -379,6 +388,7 @@ class OrganizationAuthorization(models.Model):
         return self.level in [
             OrganizationLevelRole.ADMIN.value,
             OrganizationLevelRole.SUPPORT.value,
+            OrganizationLevelRole.MARKETING.value,
         ]
 
     @property
@@ -398,6 +408,7 @@ class OrganizationAuthorization(models.Model):
             OrganizationLevelRole.ADMIN.value,
             OrganizationLevelRole.FINANCIAL.value,
             OrganizationLevelRole.SUPPORT.value,
+            OrganizationLevelRole.MARKETING.value,
         ]
 
     @property
@@ -564,6 +575,13 @@ class Project(models.Model):
         default=ProjectStatus.ACTIVE,
         max_length=8,
     )
+    language = models.CharField(
+        _("Project Language"),
+        max_length=64,
+        null=True,
+        choices=settings.LANGUAGES,
+        default=settings.DEFAULT_LANGUAGE,
+    )
 
     def __str__(self):
         return f"{self.uuid} - Project: {self.name} - Org: {self.organization.name}"
@@ -607,11 +625,6 @@ class Project(models.Model):
 
         return {"flow": flows_result, "intelligence": intelligence_result}
 
-    def perform_destroy_flows_project(self, user_email: str):
-        project_uuid = self.uuid
-
-        current_app.send_task("delete_project", args=[project_uuid, user_email])
-
     def create_classifier(self, authorization, template_type: str, access_token: str):
         flow_instance = FlowsRESTClient()
         created = False
@@ -654,8 +667,9 @@ class Project(models.Model):
         return created, data
 
     def create_chats_project(self):
-        from connect.api.v1.internal.chats.chats_rest_client import \
-            ChatsRESTClient  # to avoid circular import
+        from connect.api.v1.internal.chats.chats_rest_client import (
+            ChatsRESTClient,
+        )  # to avoid circular import
 
         created = False
         chats_client = ChatsRESTClient()
@@ -740,8 +754,7 @@ class Project(models.Model):
 
     def get_contacts(self, before: str, after: str, counting_method: str = None):
         from connect.billing.models import Contact
-        from connect.billing.utils import (custom_get_attendances,
-                                           get_attendances)
+        from connect.billing.utils import custom_get_attendances, get_attendances
 
         if not counting_method:
             counting_method = self.organization.organization_billing.plan_method
@@ -872,11 +885,11 @@ class ChatsAuthorization(models.Model):
 
 
 class ProjectRole(Enum):
-    NOT_SETTED, VIEWER, CONTRIBUTOR, MODERATOR, SUPPORT, CHAT_USER = list(range(6))
+    NOT_SETTED, VIEWER, CONTRIBUTOR, MODERATOR, SUPPORT, CHAT_USER, MARKETING = list(range(7))
 
 
 class ProjectRoleLevel(Enum):
-    NOTHING, VIEWER, CONTRIBUTOR, MODERATOR, SUPPORT, CHAT_USER = list(range(6))
+    NOTHING, VIEWER, CONTRIBUTOR, MODERATOR, SUPPORT, CHAT_USER, MARKETING = list(range(7))
 
 
 class ProjectAuthorization(models.Model):
@@ -890,6 +903,7 @@ class ProjectAuthorization(models.Model):
         (ProjectRole.MODERATOR.value, _("moderator")),
         (ProjectRole.SUPPORT.value, _("support")),
         (ProjectRole.CHAT_USER.value, _("Chat user")),
+        (ProjectRole.MARKETING.value, _("marketing")),
     ]
     uuid = models.UUIDField(
         _("UUID"), primary_key=True, default=uuid4.uuid4, editable=False
@@ -927,12 +941,15 @@ class ProjectAuthorization(models.Model):
             return ProjectRoleLevel.VIEWER.value
         elif self.role == ProjectRole.SUPPORT.value:
             return ProjectRoleLevel.SUPPORT.value
+        elif self.role == ProjectRole.MARKETING.value:
+            return ProjectRoleLevel.MARKETING.value
 
     @property
     def is_moderator(self):
         return self.level in [
             ProjectRoleLevel.MODERATOR.value,
             ProjectRoleLevel.SUPPORT.value,
+            ProjectRoleLevel.MARKETING.value,
         ]
 
     @property
@@ -940,6 +957,7 @@ class ProjectAuthorization(models.Model):
         return self.level in [
             ProjectRoleLevel.MODERATOR.value,
             ProjectRoleLevel.SUPPORT.value,
+            ProjectRoleLevel.MARKETING.value,
         ]
 
     @property
@@ -950,6 +968,7 @@ class ProjectAuthorization(models.Model):
             ProjectRoleLevel.VIEWER.value,
             ProjectRoleLevel.SUPPORT.value,
             ProjectRoleLevel.CHAT_USER.value,
+            ProjectRoleLevel.MARKETING.value,
         ]
 
     @property
@@ -958,6 +977,7 @@ class ProjectAuthorization(models.Model):
             ProjectRoleLevel.MODERATOR.value,
             ProjectRoleLevel.CONTRIBUTOR.value,
             ProjectRoleLevel.SUPPORT.value,
+            ProjectRoleLevel.MARKETING.value,
         ]