+++ connect/usecases/project/__init__.py
@@ -1,0 +1,4 @@
+from connect.usecases.project.eda_publisher import ProjectEDAPublisher
+from connect.usecases.project.update_project import UpdateProjectUseCase
+
+__all__ = ["ProjectEDAPublisher", "UpdateProjectUseCase"]